Tray: Move to src/main
This makes it match other scripts for the main process.
