+++ src/main/tray.ts
